requirements.txt
praw
pyyaml
matplotlib
matplotlib-label-lines
imgurpython
